+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyr (development version)
+
 # tidyr 1.2.0
 
 ## Pivoting